llm_project/gpt/backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-limiter
python-dotenv
openai
requests
psycopg2-binary
Flask-SocketIO>=5.3.6
eventlet>=0.33.3
python-socketio>=5.11.2

